Task #2: Implement Strategy Patt. for Different Nucleotide Calculations
Create an abstract base class for enforcing the need to have a method
that calculates the bias of the observed and expected values. This
allows for different calculations to occur on each pair of values, which
is typically the smallest unit to operate on in Excel sheets.

Verification: Run nosetests . in the directory.