Task #5: Implement Reader for Substitution Matrix with Sheet Validator
Implement a reader that specifically searches for raw data containing
the properties of a substitution matrix.  The defaults are set according
to the Input and Output files found in the data folder, which are
2SubstitutionAnlysML_INPUT.xlsx and 3normalizedSubBias_OUTPUT.xlsx.
Reader validates structure of the raw data before converting it to the
corresponding substitution matrix data structure.

Verification: Run nosetests3 . in the directory.