Fasta Renamer Using BioPython
Uses BioPython to retrieve sequence header and data to reformat the header
using only the second accession number and virus name e.g. Tomato Curly Leaf
virus to put it in the form: {accession number}_{virus name} with all spaces
replaced by underscores.  It leaves the sequence data untouched, and is just a
simple copy-paste of the data.

Verification:
 1. Edit the input_path to point to an existing FASTA formatted file.
 2. Edit the output_path to contain a validly named file.
 3. Run with: python FASTA_Renamer_BioPython.py