FASTA Renamer Using Python
Uses vanilla Python to retrieve sequence header and data to reformat the header
using only the second accession number and virus name e.g. Tomato Curly Leaf
virus, to put it in the form: {accession number}_{virus name} with all spaces
replaced by underscores.  It leaves the sequence data untouched, and is just a
simple copy-paste of the data.

This operation runs in the O(m)-space versus the O(m^2) space of the
FASTA_Renamer_BioPython module.  It performs its operations line-by-line rather
than letter-by-letter since it reads in lines from the original file in the
same fashion.  This allows for the simple use of regular expressions to
validate and select out key parts of the strings, and to just move lines over,
if no processing is required; this is the case when it comes to processing the
sequence data and where it differs from the BioPython version of this script.

Verification:
 1. Edit the input_path to point to an existing FASTA formatted file.
 2. Edit the output_path to contain a validly named file.
 3. Run with: python FASTA_Renamer_Python.py